Improve generation and automatic mode
Now the comment is inserted directly in the buffer without starting a
comment, and is commented using `comment-region` at the end.

Previously this was not done, because I did not know you could change
`comment-style` to create nice looking multiline comments. Now it is set
to `extra-line` at the start of the function to use the correct comment
style.

The inserted header is now filled automagically with the git repository
name as the project name, and the current file name as the description.
The old mode which asks for the project name and description is still
available by using the universal argument.